Create End of trial 2
This is the added changes. It includes the different colour texts in the buttons, the expanding of the saved details box, the readjustment sizing of the receipt number field, and the error messages for in case the receipt number is less than 4 numbers long or if letters have been inputed in the number of items hired field instead of numbers.