Save details columns
I have removed the listbox where the details were previously saved, now when the save details button is clicked the saved details will show up in a box with columns. This helps to make it look more organised. (Aesthetics) 
I have also slightly increased the window size vertically because with the increased box size, the quit button wasn't visible unless the user manually increased the window size, but now it will be the proper size automatically.